feat: add refer and Earn menu item
